05_archive/DISCORD-ELITEALGOBOT_BIG_FILE/requirements.txt
# Core Dependencies
python-dotenv>=0.19.0
discum>=1.3.0
ccxt>=2.0.0

# Development
pytest>=6.2.5
pytest-asyncio>=0.15.1
black>=21.9b0
isort>=5.9.3
flake8>=3.9.2
mypy>=0.910

# Logging
colorlog>=6.6.0

# Utilities
python-dateutil>=2.8.2
pytz>=2021.3
